feat: add game editing workflow

+++ src/components/GameCard.ts
@@ -1,31 +1,50 @@
 import type { Game } from "../types/Game";
 import { GameService } from "../services/GameService.js";
-import { initGameCardList } from "./GameCardList.js";
+import { initGameCardList, selectGame, clearSelection } from "./GameCardList.js";
 
-// GameCard ===============================================================
 export class GameCard {
 
-    static create(game: Game): HTMLElement {
+    static create(game: Game, selected: boolean): HTMLElement {
 
         const card = document.createElement("article");
         card.className = "card game-card";
-        //image
+
+        //select gamecard
+        if (selected) {
+            card.classList.add("border-accent");
+        }
+
+        // Select / deselect card
+        card.addEventListener("click", () => {
+            selectGame(game.id);
+        });
+
+        // image
         const image = document.createElement("img");
         image.src = game.image;
         image.alt = game.name;
-        //title
+
+        // title
         const title = document.createElement("h3");
         title.textContent = game.name;
-        //delete button 
+
+        // delete button
         const buttonRemove = document.createElement("button");
         buttonRemove.type = "button";
         buttonRemove.textContent = "×";
-        buttonRemove.addEventListener("click", () => {
+        buttonRemove.addEventListener("click", event => {
+            event.stopPropagation();
+            if (!confirm(`Deseja excluir "${game.name}"?`)) {
+                return;
+            }
             GameService.remove(game.id);
+            clearSelection();
             initGameCardList();
         });
 
+        //append
         card.append(image, title, buttonRemove);
+
         return card;
     }
 }
